pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flux2-resolution-guard"
version = "0.1.0"
description = "FLUX.2 high-resolution drift correction and ComfyUI node pack"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "OpenAI" }
]
dependencies = [
  "torch>=2.1",
  "Pillow>=10.0",
  "numpy>=1.26",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
